Правки модуля5
