+++ js/data.js
@@ -1,10 +1,8 @@
 import {getRandomArrayElement} from './utils.js';
 import {getRandomPositiveInteger} from './utils.js';
 import {getRandomPositiveFloat} from './utils.js';
-import {getRandomFeatures} from './main.js';
-import {makeGenerator} from './main.js';
-import {getAvatar} from './main.js';
 import {getRandomFeatures} from './utils.js';
+import {getAvatar} from './utils.js';
 
 const offerTitle = [
   'Добро пожаловать в наш шикарный дворец!',
